+++ library/core/src/main/java/com/google/android/exoplayer2/text/webvtt/WebvttCssStyle.java
@@ -263,7 +263,7 @@ public final class WebvttCssStyle {
     return this;
   }
 
-  public WebvttCssStyle setFontSizeUnit(short unit) {
+  public WebvttCssStyle setFontSizeUnit(@FontSizeUnit int unit) {
     this.fontSizeUnit = unit;
     return this;
   }